+++ modules/app-desktop/src/main/kotlin/org/archivekeep/app/desktop/ui/views/storages/components/AllStoragesList.kt
@@ -1,16 +1,12 @@
 package org.archivekeep.app.desktop.ui.views.storages.components
 
 import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.ExperimentalLayoutApi
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.material.Icon
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
@@ -20,27 +16,25 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.cheonjaeung.compose.grid.SimpleGridCells
 import com.cheonjaeung.compose.grid.VerticalGrid
-import compose.icons.TablerIcons
-import compose.icons.tablericons.DotsVertical
 import org.archivekeep.app.desktop.ui.components.LoadableGuard
 import org.archivekeep.app.desktop.ui.components.richcomponents.StorageDropdownIconLaunched
 import org.archivekeep.app.desktop.ui.designsystem.sections.SectionCard
+import org.archivekeep.app.desktop.ui.designsystem.sections.SectionCardActionsRow
 import org.archivekeep.app.desktop.ui.designsystem.sections.SectionCardBottomList
 import org.archivekeep.app.desktop.ui.designsystem.sections.SectionCardTitle
 import org.archivekeep.app.desktop.ui.designsystem.sections.sectionCardHorizontalPadding
 import org.archivekeep.app.desktop.ui.views.storages.StoragesVM
 import org.archivekeep.utils.loading.Loadable
 
-@OptIn(ExperimentalLayoutApi::class)
 @Composable
-fun AllStoragesList(allStorages: Loadable<List<StoragesVM.Storage>>) {
-    LoadableGuard(allStorages) { allLocalArchives ->
+fun AllStoragesList(allStoragesLoadable: Loadable<List<StoragesVM.Storage>>) {
+    LoadableGuard(allStoragesLoadable) { allStorages ->
         VerticalGrid(
             columns = SimpleGridCells.Adaptive(minSize = 240.dp),
             horizontalArrangement = Arrangement.spacedBy(8.dp),
             verticalArrangement = Arrangement.spacedBy(8.dp),
         ) {
-            allLocalArchives.forEach { storage ->
+            allStorages.forEach { storage ->
                 SectionCard {
                     SectionCardTitle(
                         // TODO
@@ -49,6 +43,11 @@ fun AllStoragesList(allStorages: Loadable<List<StoragesVM.Storage>>) {
                         icons = {
                             StorageDropdownIconLaunched(storage.uri)
                         },
+                    )
+
+                    SectionCardActionsRow(
+                        emptyList(),
+                        noActionsText = if (storage.isLocal) "Local storage" else "External storage",
                     )
 
                     Spacer(Modifier.height(4.dp))
@@ -77,18 +76,6 @@ fun AllStoragesList(allStorages: Loadable<List<StoragesVM.Storage>>) {
                                     fontSize = 14.sp,
                                     lineHeight = 16.sp,
                                 )
-                            }
-
-                            Row(
-                                verticalAlignment = Alignment.CenterVertically,
-                            ) {
-                                Box(Modifier.padding(6.dp)) {
-                                    Icon(
-                                        TablerIcons.DotsVertical,
-                                        contentDescription = "Upload",
-                                        Modifier.size(16.dp),
-                                    )
-                                }
                             }
                         }
                     }